merge in switch-to-cxv branch
(This doesn't actually switch to CXF, though; I determined that switching would not solve the load-balanced-url problem, so instead I stuck with the jax-ws RI and found a different solution)

The major goal of this branch was to fix a problem that has kept us from using a load-balanced https url to Answers instead of using the direct http appserver url.  The previous configuration pulled down the wsdl hosted by answers, which had 2 problems:
-it had a bug in it (using wsdl:import instead of xsd:import to pull in the schema file)
-it referenced http endpoint urls instead of https urls, and the jax-ws client can't handle 302 http->https redirects

We now keep a fixed-up wsdl in source control, and generate code from that.  In addition, we use this wsdl at runtime, as well.  This solves another problem we've had: if Answers is down when someone tries to log into Staff Services, they get an error screen (even if they're not trying to go to a page backed by Answers data).  

This change alters the obiee.properties format - instead of obiee.wsdl.url, obiee.endpoint.urlBase should be specified.  Since a bunch of obiee.properies files will need to be changed anyway (to use the load-balanced urls), I deemed this a reasonable change.






git-svn-id: http://arkham.ccci.org/svn/java/trunk/commons/obiee-ws-client@10928 6185d61c-cf3b-044e-8ca2-f4aa9eb47c87