Set JDK jax-ws timeouts
The previous code assumed the client is either using CXF or the jax-ws RI,
and not the jax-ws impl that is built into the sun/oracle jdk.
Our Give/EP project uses the built-in jdk jax-ws impl
(for better or worse).

Note that I could use the constants defined on
com.sun.xml.internal.ws.client.BindingProviderProperties
but it would require extra work to have Maven allow us
to use a jdk implementation class.
This seems easier, and I am not too worried about the duplication in this case.