Add OpenTracing
Also, remove the pseudo-APM OperationTimer, which was really only used for tests before.
OpenTracing is now used for tests and at runtime.

This commit also upgrades the project to Java 8
(which is long overdue).
Mainly, it was because Stopwatch was marked @Beta and intellij was giving me a warning for it.
At first I upgraded Guava to the latest, which required upgrading to Java 8.
But it turns out Elastic Path (16.x) cannot use guava >= 19
(at least not without some probably-esoteric work),
so I downgraded to guava 18.

Note: this commit uses opentracing 0.31.0,
because this is the version that DataDog uses,
but the latest release candidate for 0.32.0 is out and it deprecates some methods.
I've avoided using some of those deprecated methods in this PR,
but some cannot be avoided until we use replacements introduced in 0.32.0.

This commit also moves the slf4j compile dependency up
with the other compile-scoped dependencies.