Record errors on top-level spans
To get obiee-specific error rates in DataDog,
we need to record errors on the spans.
We only need to record these on top-level spans,
since the lower-level spans here always throw exceptions on errors,
and don't swallow them.

A few top-level spans don't need to record errors,
because they're just used for validation,
and errors in validation queries can still result in successful requests
(in a new session).