Record some context details as fields on our exceptions
In particular, I want to have the report name recorded as a field.
I am working towards setting up rollbar fingerprinting rules that will
group socket read timeout errors together for each report.
(I don't want all read timeouts grouped together, since read timeouts
can be isolated to a particular query's design/performance.)
But to have rollbar rules be able to use the report name,
it needs to be available as an individual attribute in the report json.
Red-egg almost does this.
I'll be adding logic to complete this there, soon.

While doing this, I'm moving some other context info to fields, as well.
As is, the 'params' chunk of the exception message can be obnoxiously big,
and it's better to put it into an exception field.
Same with the fault detail string.